reorg api-related content; remove unused view
